app/requirements.txt
fastapi
uvicorn
psycopg2-binary
faker
requests
pytest
